+++ AngularClient/src/main/core/project.service.ts
@@ -193,22 +193,7 @@ export class ProjectService {
 
     return this.appEntityManager.executeQueryObservable<Project>(query, forceRefresh).pipe(
       map(response => {
-        var project = response.results[0] || null;
-        console.log("p", project);
-
-        project.ElementSet.forEach(element => {
-          element.ElementFieldSet.forEach(field => {
-
-            // Default 
-            field.ElementCellSet.forEach(cell => {
-              console.log("c", cell.DecimalValueTotal, cell.DecimalValueCount, cell.UserElementCellSet.length);
-            });
-          });
-        });
-
-        console.log(" --- ");
-
-        return project;
+        return response.results[0] || null;
       }));
   }